refactor: move KycOutboxJobSettings to kyc/config.rs (#10)
Mirrors the indexer module pattern where job settings live in
their own config.rs file within the feature directory.
