+++ packages/worker/src/kyc/mod.rs
@@ -1,1 +1,2 @@
+pub mod config;
 pub mod kyc_outbox;
